test/cluster: move commitlog crash-recovery tests to separate file
Move the 6 commitlog crash-recovery tests from test_strong_consistency.py
to a new test_strong_consistency_commitlog.py for better organization.

As the test_strong_consistency.py file grows, grouping closely related
tests into separate files makes navigation easier, simplifies finding
tests for a specific feature, and allows running only the relevant
subset.
